requirements.txt
amqp==2.6.1
asgiref==3.3.4
astroid==2.5.6
billiard==3.6.4.0
celery==4.4.7
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.2.2
django-celery-results==2.0.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
dnspython==1.16.0
flower==0.9.7
gevent==21.1.2
greenlet==1.1.0
humanize==3.5.0
isort==5.8.0
kombu==4.6.11
lazy-object-proxy==1.6.0
mccabe==0.6.1
pika==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycparser==2.20
pylint==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
vine==1.3.0
wcwidth==0.2.5
windows-curses==2.2.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
